Move `ftplugin` to `plugin`
Make this a regular plugin instead of a filetype plugin. Rails locale
files are YAML files, but we don't want these commands to be available
for all YAML files, only those that match the path for Rails locales.
We won't be able to do the path match and filetype separation with an
ftplugin, so instead we'll set up the commands to be defined by
`autocmd`s that match the Rails locale path.